Pay and payment confirmation

+++ src/main/java/com/solar/ms/rms/model/payment/ReservePaymentRequest.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 @Data
 public class ReservePaymentRequest {
     private String productName;
-    private BigDecimal amount;
+    private double amount;
     private String orderId;
     private String confirmUrl;
 }
